requirements.txt
aiogram==3.4.1
SQLAlchemy==2.0.30
python-dotenv==1.0.1
